feat: Add weapon collision

+++ Ball Battle Roguelike/Assets/Scripts/CharacterController.cs
@@ -10,8 +10,8 @@ public class CharacterController : MonoBehaviour, IDamageable
     //CircleCollider2D collider2d;
     LayerMask collisionMask;
 
-    [SerializeField] int health = 10;   // The amount of damage the character can take before dying
-    [SerializeField] int speed = 0;     // The speed the character moves at
+    [SerializeField] double health = 10;   // The amount of damage the character can take before dying
+    [SerializeField] double speed = 0;     // The speed the character moves at
 
     Vector2 moveDirection;
 
@@ -28,7 +28,7 @@ public class CharacterController : MonoBehaviour, IDamageable
 
     private void FixedUpdate()
     {
-        float distance = speed * Time.fixedDeltaTime;
+        float distance = (float)speed * Time.fixedDeltaTime;
         while (distance > 0)
         {
             RaycastHit2D hit = Physics2D.CircleCast(
@@ -54,8 +54,8 @@ public class CharacterController : MonoBehaviour, IDamageable
         }
     }
 
-    public void TakeDamage(int damage, DamageType damageType)
+    public void TakeDamage(double damage, DamageType damageType)
     {
-        health -= Mathf.Max(0, damage);
+        health -= Mathf.Max(0, (float)damage);
     }
 }
